revisi prioritas dan galeri

+++ admin/galeri/hapus.php
@@ -6,9 +6,18 @@ $id = $_GET['id'];
 $queryDetailPaket = mysqli_query($koneksi,"SELECT gambar FROM galeri WHERE id = '$id'");
 $d = mysqli_fetch_assoc($queryDetailPaket);
 $gambar = $d['gambar'];
+
 $queryDelete = "DELETE FROM galeri WHERE id = '$id'";
 mysqli_query($koneksi,$queryDelete);
-unlink('../../'.$gambar);
+
+$queryDelete = "DELETE FROM galeri_detail WHERE galeri_id = '$id'";
+mysqli_query($koneksi,$queryDelete);
+
+if($gambar){
+	unlink('../../'.$gambar);
+}
+
+
 echo "<script>
 	alert('Berhasil di hapus!');
 	window.location.href='index';
